+++ packages/cli/src/__tests__/e2e/test.spec.ts
