+++ packages/cli/src/commands/index.ts
@@ -3,6 +3,6 @@ export * from "./codegen";
 export * from "./create";
 export * from "./deploy";
 export * from "./infra";
-export * from "./run";
+export * from "./test";
 export * from "./docgen";
 export * from "./manifest";
