+++ packages/cli/src/commands/test.ts
